+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - CLI Options Renamed
   - `--depends-on-components` renamed to `--modified-components`
   - `--depends-on-files` renamed to `--modified-files`
+  - `--ignore-components-dependencies-file-patterns` renamed to `--ignore-app-dependencies-filepatterns`
 - Removed the deprecated CLI call methods, now these options only support space-separated list
   - `--exclude`
   - `--config`